fix: restructure existing tests
