+++ test/integration/server_response/test_GET.py
@@ -29,14 +29,6 @@ def test_GET_directory_listing():
     assert heading in response.text
     assert response.status_code == 200
 
-def test_GET_directory_redirect():
-    print("Request for /directory")
-    url = "http://localhost:8080/directory"
-    response = make_request(url)
-# does not work right now
-#    assert response.headers["location"] == "/directory/"
-    assert response.status_code == 200
-
 def test_GET_percent_encoded():
     print("Request for /images/grüne äpfel.jpg")
     url ="http://localhost:8080/images/grüne äpfel.jpg"
@@ -51,45 +43,6 @@ def test_GET_location_with_alias():
     file_size = os.path.getsize(file_path)
     assert int(response.headers["content-length"]) == file_size
     assert response.status_code == 200
-
-def test_GET_directory_forbidden():
-    print("Request for /secret/")
-    url = "http://localhost:8080/secret/"
-    response = make_request(url)
-    assert response.status_code == 403
-
-def test_GET_simple_redirect():
-    print("Request for /redirect")
-    url = "http://localhost:8080/redirect"
-    response = make_request(url)
-    assert response.history[0].status_code == 301
-    assert response.history[0].headers["Location"] == "/secret"
-    assert response.status_code == 403
-
-def test_GET_return_forbidden_error_page_returns_also():
-    print("Request for /another")
-    url = "http://localhost:8080/another"
-    response = make_request(url)
-    assert response.status_code == 403
-    assert "Return Message" in response.text
-
-def test_GET_missing_indices():
-    print("Request for /missingIndex/")
-    url = "http://localhost:8080/missingIndex/"
-    response = make_request(url)
-    assert response.status_code == 403
-
-def test_max_recurision():
-    print("Request for /recursion/")
-    url = "http://localhost:8080/recursion/"
-    response = make_request(url)
-    assert response.status_code == 500
-
-def test_character_device():
-    print("Request for /tty")
-    url = "http://localhost:8080/tty"
-    response = make_request(url)
-    assert response.status_code == 500
 
 def test_GET_empty_return_directive():
     print("Request for /empty")
